It is working now

+++ 4+/1/main.c
@@ -26,7 +26,7 @@ char bit = 0;
 
 void hndlSIGUSR(int signo)
 {
-	fprintf(stderr, "IMSIG\n");
+	//fprintf(stderr, "IMSIG\n");
 	switch(signo)
 	{
 		case SIGUSR1:
@@ -40,6 +40,11 @@ void hndlSIGUSR(int signo)
 	}
 }
 
+void hndl(int signo)
+{
+	//fprintf(stderr, "HI\n");
+	return;
+}
 
 int StartClient(pid_t pid)
 {
@@ -47,72 +52,74 @@ int StartClient(pid_t pid)
 	struct sigaction act = {};
 	struct sigaction oact = {};
 	char byte = 0;
-
-	_(sigemptyset(&set));
 	_(sigaddset(&set, SIGUSR1));
 	_(sigaddset(&set, SIGUSR2));
-	PL;
-	//(sigprocmask(SIG_BLOCK, &set, NULL));
+
 	SIGACTION(SIGUSR1, hndlSIGUSR, set, 0);
 	SIGACTION(SIGUSR2, hndlSIGUSR, set, 0);
-	PL;
+	SIGACTION(SIGCHLD, hndlSIGUSR, set, 0);
+	exit(0);
+	_(sigemptyset(&set));
 	//sleep(1);
 	while(1)
 	{
 		for (int i = 0; i < 8; i++)
 		{
-			PL;
-			_(sigsuspend(&set));
-			PL;
-			_(bit);
-			byte = (byte << 1) & bit;
+			sigsuspend(&set);
+			if (bit < 0)
+				return 0;
+			byte = (byte << 1) | bit;
 			_(kill(pid, SIGUSR1));
 		}
-		_(write(1, &byte, sizeof(byte)));
+		_(write(STDOUT_FILENO, &byte, sizeof(byte)));
+		//fflush(stdin);
 	}
-
 	return 0;
 }
 
 int StartServer(const char* name)
 {
-	PL;
 	sigset_t set = {};
+	struct sigaction act = {};
+	struct sigaction oact = {};
 	pid_t ppid = getppid();
 	char byte = 0;
 	size_t nread = 0;
 	int fd = 0;
+
+
+
 	_(fd = open(name, O_RDONLY));
-	//sleep(1);
+	// PL;
 	_(sigemptyset(&set));
-	_(sigaddset(&set, SIGUSR1));
-	//sleep(1);
-	PL;
+	SIGACTION(SIGUSR1, hndl, set, 0);
+	// PL;
+	
 	do
 	{
 		_(nread = read(fd, &byte, sizeof(byte)));
 		for (int i = 0; i < 8 * sizeof(byte); i++)
 		{
-			PL;
 			switch(byte & (1 << 8))
 			{
 				case 0:
 					kill(ppid, SIGUSR1);
 					break;
-				case 1:
+				case (1 << 8):
 					kill(ppid, SIGUSR2);
 					break;
 				default:
 					return -1;
 			}
-			PL;
 			byte = byte << 1;
-			_(sigsuspend(&set));
-			PL;
+			_(sigemptyset(&set));
+
+			sigsuspend(&set);
+
 		}
 	}while(nread);
 
-	_(kill(ppid, SIGKILL));
+	//_(kill(ppid, SIGKILL));
 
 	close(fd);
 	return 0;
@@ -127,7 +134,12 @@ int main(int argc, char* argv[])
 		printf("Use: %s <filename>\n", argv[0]);
 		return -1;
 	}
-	
+	sigset_t set = {};
+	_(sigemptyset(&set));
+	_(sigaddset(&set, SIGUSR1));
+	_(sigaddset(&set, SIGUSR2));
+	_(sigprocmask(SIG_BLOCK, &set, NULL));
+	errno = 0;
 	pid_t pid = fork();
 	if ( pid > 0 )
 	{
@@ -147,7 +159,7 @@ int main(int argc, char* argv[])
 	}
 	else
 	{
-		printf("Error in fork()\n");
+		perror("Error in fork()\n");
 		return -1;
 	}
 	return 0;
